+++ docs/api.rst
@@ -124,7 +124,6 @@ Inference
    test_statistics.qmu_tilde
    test_statistics.tmu
    test_statistics.tmu_tilde
-   test_statistics.get_test_stat
    mle.twice_nll
    mle.fit
    mle.fixed_poi_fit
@@ -136,6 +135,7 @@ Inference
    calculators.AsymptoticCalculator
    calculators.ToyCalculator
    utils.create_calculator
+   utils.get_test_stat
 
 Exceptions
 ----------
